test(unittests): consolidate and refactor test fixtures for starlark
- Moved common test fixtures (`mock_tavern_internal`, `mock_test_config`, `mock_response`, `basic_runner`) to a new shared file: `tests/unit/starlark/conftest.py`, enabling reuse across test suites.
- Removed duplicate fixture definitions from `test_starlark_env.py` and `test_response_struct.py` to reduce redundancy.
- Simplified the test structure by centralizing reusable components, improving maintainability and clarity of test files.
